+++ src/api/place.api.ts
@@ -2,11 +2,12 @@ import { Hono } from "hono";
 import { prisma } from '../prisma.js'
 import {zValidator} from '@hono/zod-validator'
 import {pagingSchema,createPlaceSchema,updatePlaceSchema} from "../schema.zod.js"
+import {authenticateAdmin, authenticate} from "../authentication/jwtauth.js"
 
 export const placeApi = new Hono();
 
 //ná í 
-placeApi.get('/',zValidator('query',pagingSchema) ,async(c)=>{
+placeApi.get('/',authenticate,zValidator('query',pagingSchema) ,async(c)=>{
     const limit=c.req.valid('query').limit
     const offset =c.req.valid('query').offset
 
@@ -27,7 +28,7 @@ placeApi.get('/',zValidator('query',pagingSchema) ,async(c)=>{
 })
 
 //Ná í eftir id eða slug
-placeApi.get('/:id',zValidator('query',pagingSchema) ,async(c)=>{
+placeApi.get('/:id',authenticate,zValidator('query',pagingSchema) ,async(c)=>{
     const id = c.req.param('id')
 
     const place = await prisma.place.findUnique({
@@ -47,7 +48,7 @@ placeApi.get('/:id',zValidator('query',pagingSchema) ,async(c)=>{
 })
 
 //Búa til
-placeApi.post('/',zValidator('query',createPlaceSchema,(result, c) => { if (!result.success) {
+placeApi.post('/',authenticateAdmin,zValidator('query',createPlaceSchema,(result, c) => { if (!result.success) {
       return c.json("Bad request",400)}}), async(c)=>{
         const email=c.req.valid('query').email
         const address =c.req.valid('query').address
@@ -71,7 +72,7 @@ placeApi.post('/',zValidator('query',createPlaceSchema,(result, c) => { if (!res
 
 
 //Uppfæra
-placeApi.put('/:id',zValidator('query',updatePlaceSchema,(result, c) => {if (!result.success) {
+placeApi.put('/:id',authenticateAdmin,zValidator('query',updatePlaceSchema,(result, c) => {if (!result.success) {
     return c.json("Bad request",400)}}), async(c)=>{
         const id = c.req.param('id')
         const email=c.req.valid('query').email
@@ -99,10 +100,10 @@ placeApi.put('/:id',zValidator('query',updatePlaceSchema,(result, c) => {if (!re
     
 
 //Eyða
-placeApi.delete('/:id',zValidator('query',pagingSchema) ,async(c)=>{
+placeApi.delete('/:id',authenticateAdmin,zValidator('query',pagingSchema) ,async(c)=>{
     const id = c.req.param('id')
 
-    await prisma.event.delete({
+    await prisma.image.delete({
     where: {
         id:Number(id),},});
